+++ plugin-shuiyin1/src/client/index.tsx
@@ -1,0 +1,3 @@
+import { PluginShuiyin1Client } from './plugin';
+
+export default PluginShuiyin1Client;