+++ plugin-shuiyin1/src/server/index.ts
@@ -1,0 +1,3 @@
+import { PluginShuiyin1Server } from './plugin';
+
+export default PluginShuiyin1Server;